whole lot of logans
- move generic DNS stuff from smithdnsupd.* to dns.*

- move functions from main.c to smithproxy.cpp, SmithProxy class
