+++ dns.hpp
@@ -58,6 +58,8 @@
 #include <ext/cidr.hpp>
 #include <addrobj.hpp>
 
+#include <dns.hpp>
+
 #define DNS_HEADER_SZ 12
 
 
@@ -71,6 +73,8 @@ typedef enum DNS_Record_Type_ {
     OPT=41
 } DNS_Record_Type;
 
+
+struct DNS_Response;
 
 class DNSFactory {
     const char* _unknown = "unknown";
@@ -92,6 +96,20 @@ public:
     const char *dns_record_type_str(int a);
     int load_qname(unsigned char* ptr, unsigned int maxlen, std::string* str_storage);
     int generate_dns_request(unsigned short id, buffer& b, std::string const& hostname, DNS_Record_Type t);
+
+    // send DNS request out to network. Return socket FD, or non-positive on error.
+    // you want to call this for async request
+    int send_dns_request (std::string const &hostname, DNS_Record_Type t, std::string const &nameserver);
+
+    // @returns: response and return from receive - well. I don't like it this way either,
+    // but we can't actually return nullptr, since it could be legit return value on non-blocking socket
+    // you want to call this for async receive - ideally if socket is in readset.
+    std::pair<DNS_Response*,int> recv_dns_response(int send_socket, unsigned int timeout_sec=2);
+
+
+    // this is easiest way to resolve. Just do the thing, with blocking ... and waiting.
+    DNS_Response* resolve_dns_s (std::string const& hostname, DNS_Record_Type t, std::string const& nameserver, unsigned int timeout_s=2);
+
 
     static DNSFactory& get() {
         static DNSFactory d;
@@ -209,7 +227,7 @@ private:
 
 public:
     explicit DNS_Packet() {
-        log = logan::attach<DNS_Packet>(this, "dns");
+        log = logan::attach<DNS_Packet>(this, "com.dns");
     };
 
     std::vector<int> answer_ttl_idx; // should be protected;
