+++ main.cpp
@@ -95,314 +95,7 @@
 #include <smithproxy.hpp>
 
 
-volatile static int cnt_terminate = 0;
-static bool cfg_daemonize = false;
 
-#ifndef MEM_DEBUG
-bool cfg_openssl_mem_dbg = false;
-bool cfg_mtrace_enable = false;
-#else
-bool cfg_openssl_mem_dbg = true;
-static bool cfg_mtrace_enable = true;
-#endif
-
-
-
-
-void my_terminate (int param) {
-
-    if (!cfg_daemonize)
-        ERRS_("Terminating ...\n");
-
-    SmithProxy::instance().stop();
-
-    cnt_terminate++;
-    if(cnt_terminate == 3) {
-        if (!cfg_daemonize)
-            FATS_("Failed to terminate gracefully. Next attempt will be enforced.\n");
-    }
-    if(cnt_terminate > 3) {
-        if (!cfg_daemonize)
-            FATS_("Enforced exit.\n");
-        abort();
-    }
-}
-
-
-void my_usr1 (int param) {
-    DIAS_("USR1 signal handler started");
-    NOTS_("reloading policies and its objects !!");
-    load_config(CfgFactory::get().config_file,true);
-    DIAS_("USR1 signal handler finished");
-}
-
-
-
-
-int load_signatures(libconfig::Config& cfg, const char* name, std::vector<duplexFlowMatch*>& target) {
-    using namespace libconfig;
-    
-    const Setting& root = cfg.getRoot();
-    const Setting& cfg_signatures = root[name];
-    int sigs_len = cfg_signatures.getLength();
-
-    
-    DIA_("Loading %s: %d",name,sigs_len);
-    for ( int i = 0 ; i < sigs_len; i++) {
-        MyDuplexFlowMatch* newsig = new MyDuplexFlowMatch();
-        
-        
-        const Setting& signature = cfg_signatures[i];
-        signature.lookupValue("name", newsig->name());
-        signature.lookupValue("side", newsig->sig_side);
-        signature.lookupValue("cat", newsig->category);
-        signature.lookupValue("severity", newsig->severity);
-
-        const Setting& signature_flow = cfg_signatures[i]["flow"];
-        int flow_count = signature_flow.getLength();
-        
-        DIA_("Loading signature '%s' with %d flow matches",newsig->name().c_str(),flow_count);
-
-        
-        for ( int j = 0; j < flow_count; j++ ) {
-
-            std::string side;
-            std::string type;
-            std::string sigtext;
-            int bytes_start;
-            int bytes_max;
-            
-            if(!(signature_flow[j].lookupValue("side", side)
-                && signature_flow[j].lookupValue("type", type)
-                && signature_flow[j].lookupValue("signature", sigtext)
-                && signature_flow[j].lookupValue("bytes_start", bytes_start)
-                && signature_flow[j].lookupValue("bytes_max", bytes_max))) {
-                
-                WAR_("Starttls signature %s failed to load: index %d",i);
-                continue;
-            }
-            
-            if( type == "regex") {
-                DEB_(" [%d]: new regex flow match",j);
-                newsig->add(side[0],new regexMatch(sigtext,bytes_start,bytes_max));
-            } else
-            if ( type == "simple") {
-                DEB_(" [%d]: new simple flow match",j);
-                newsig->add(side[0],new simpleMatch(sigtext,bytes_start,bytes_max));
-            }
-        }
-        
-        target.push_back(newsig);
-    }    
-    
-    return sigs_len;
-}
-
-bool init_syslog() {
-
-
-    // create UDP socket
-    int syslog_socket = socket(AF_INET, SOCK_DGRAM, IPPROTO_UDP); 
-
-    struct sockaddr_storage syslog_in;
-    memset(&syslog_in, 0, sizeof(struct sockaddr_storage));
-    
-    if(CfgFactory::get().syslog_family != 6) {
-        CfgFactory::get().syslog_family = 4;
-        syslog_in.ss_family                = AF_INET;
-        ((sockaddr_in*)&syslog_in)->sin_addr.s_addr = inet_addr(CfgFactory::get().syslog_server.c_str());
-        if(((sockaddr_in*)&syslog_in)->sin_addr.s_addr == INADDR_NONE) {
-            ERR_("Error initializing syslog server: %s", CfgFactory::get().syslog_server.c_str());
-            return false;
-        }
-        
-        ((sockaddr_in*)&syslog_in)->sin_port = htons(CfgFactory::get().syslog_port);
-    } else {
-        CfgFactory::get().syslog_family = 6;
-        syslog_in.ss_family                = AF_INET6;
-        int ret = inet_pton(AF_INET6, CfgFactory::get().syslog_server.c_str(),(unsigned char*)&((sockaddr_in6*)&syslog_in)->sin6_addr.s6_addr);
-        if(ret <= 0) {
-            ERR_("Error initializing syslog server: %s", CfgFactory::get().syslog_server.c_str());
-            return false;
-        }
-        ((sockaddr_in6*)&syslog_in)->sin6_port = htons(CfgFactory::get().syslog_port);
-    }
-    
-    
-    ::connect(syslog_socket,(sockaddr*)&syslog_in,sizeof(sockaddr_storage));
-    
-    get_logger()->remote_targets(string_format("syslog-udp%d-%d", CfgFactory::get().syslog_family, syslog_socket),syslog_socket);
-
-    logger_profile* lp = new logger_profile();
-    
-    lp->logger_type = logger_profile::REMOTE_SYSLOG;
-    lp->level_ = CfgFactory::get().syslog_level;
-    
-    // raising internal logging level
-    if(lp->level_ > get_logger()->level()) {
-        NOT_("Internal logging raised from %d to %d due to syslog server loglevel.",get_logger()->level(), lp->level_);
-        get_logger()->level(lp->level_);
-    }
-    
-    lp->syslog_settings.severity = lp->level_.level();
-    lp->syslog_settings.facility = CfgFactory::get().syslog_facility;
-    
-    get_logger()->target_profiles()[(uint64_t)syslog_socket] = lp;
-    
-    return true;
-}
-
-bool load_config(std::string& config_f, bool reload) {
-    bool ret = true;
-    auto this_daemon = DaemonFactory::instance();
-    
-    using namespace libconfig;
-    if(! CfgFactory::get().cfgapi_init(config_f.c_str()) ) {
-        FATS_("Unable to load config.");
-        ret = false;
-    }
-
-    CfgFactory::get().config_file = config_f;
-    
-    // Add another level of lock. File is already loaded. We need to apply its content.
-    // lock is needed here to not try to match against potentially empty/partial policy list
-    std::lock_guard<std::recursive_mutex> l_(CfgFactory::lock());
-    try {
-        
-        if(reload) {
-            CfgFactory::get().cfgapi_cleanup();
-        }
-
-        CfgFactory::get().load_db_address();
-        CfgFactory::get().load_db_port();
-        CfgFactory::get().load_db_proto();
-        CfgFactory::get().load_db_prof_detection();
-        CfgFactory::get().load_db_prof_content();
-        CfgFactory::get().load_db_prof_tls();
-        CfgFactory::get().load_db_prof_alg_dns();
-        CfgFactory::get().load_db_prof_auth();
-
-        CfgFactory::get().load_db_policy();
-        
-        
-        if(!reload)  {
-            load_signatures(CfgFactory::cfg_obj(),"detection_signatures",sigs_detection);
-            load_signatures(CfgFactory::cfg_obj(),"starttls_signatures",sigs_starttls);
-        }
-
-        CfgFactory::get().load_settings();
-
-        CfgFactory::cfg_root()["debug"].lookupValue("log_data_crc",baseCom::debug_log_data_crc);
-        CfgFactory::cfg_root()["debug"].lookupValue("log_sockets",baseHostCX::socket_in_name);
-        CfgFactory::cfg_root()["debug"].lookupValue("log_online_cx_name",baseHostCX::online_name);
-        CfgFactory::cfg_root()["debug"].lookupValue("log_srclines",get_logger()->print_srcline());
-        CfgFactory::cfg_root()["debug"].lookupValue("log_srclines_always",get_logger()->print_srcline_always());
-
-        CfgFactory::cfg_root()["debug"]["log"].lookupValue("sslcom",SSLCom::log_level_ref().level_ref());
-        CfgFactory::cfg_root()["debug"]["log"].lookupValue("sslmitmcom",baseSSLMitmCom<SSLCom>::log_level_ref().level_ref());
-        CfgFactory::cfg_root()["debug"]["log"].lookupValue("sslmitmcom",baseSSLMitmCom<DTLSCom>::log_level_ref().level_ref());
-        CfgFactory::cfg_root()["debug"]["log"].lookupValue("sslcertstore",SSLFactory::get_log().level()->level_ref());
-        CfgFactory::cfg_root()["debug"]["log"].lookupValue("proxy",baseProxy::log_level_ref().level_ref());
-        CfgFactory::cfg_root()["debug"]["log"].lookupValue("proxy",epoll::log_level.level_ref());
-        CfgFactory::cfg_root()["debug"]["log"].lookupValue("mtrace",cfg_mtrace_enable);
-        CfgFactory::cfg_root()["debug"]["log"].lookupValue("openssl_mem_dbg",cfg_openssl_mem_dbg);
-
-        /*DNS ALG EXPLICIT LOG*/
-        CfgFactory::cfg_root()["debug"]["log"].lookupValue("alg_dns",DNS_Inspector::log_level_ref().level_ref());
-        CfgFactory::cfg_root()["debug"]["log"].lookupValue("alg_dns",DNS_Packet::log_level_ref().level_ref());
-
-
-        CfgFactory::cfg_root()["settings"].lookupValue("write_payload_dir",CfgFactory::get().traflog_dir);
-        CfgFactory::cfg_root()["settings"].lookupValue("write_payload_file_prefix",CfgFactory::get().traflog_file_prefix);
-        CfgFactory::cfg_root()["settings"].lookupValue("write_payload_file_suffix",CfgFactory::get().traflog_file_suffix);
-
-
-
-        // initialize stubborn logans :)
-        auto _ = inet::Factory::log();
-
-
-        // don't mess with logging if just reloading
-        if(! reload) {
-
-            
-            //init crashlog file with dafe default
-            this_daemon.set_crashlog("/tmp/smithproxy_crash.log");
-
-            if(CfgFactory::cfg_root()["settings"].lookupValue("log_file",CfgFactory::get().log_file_base)) {
-
-                CfgFactory::get().log_file = CfgFactory::get().log_file_base;
-
-
-                if(CfgFactory::get().log_file.size() > 0) {
-
-                    CfgFactory::get().log_file = string_format(CfgFactory::get().log_file.c_str(), CfgFactory::get().tenant_name.c_str());
-                    // prepare custom crashlog file
-                    std::string crlog = CfgFactory::get().log_file + ".crashlog.log";
-                    this_daemon.set_crashlog(crlog.c_str());
-                    
-                    std::ofstream * o = new std::ofstream(CfgFactory::get().log_file.c_str(),std::ios::app);
-                    get_logger()->targets(CfgFactory::get().log_file, o);
-                    get_logger()->dup2_cout(false);
-                    get_logger()->level(cfgapi_table.logging.level);
-                    
-                    logger_profile* lp = new logger_profile();
-                    lp->print_srcline_ = get_logger()->print_srcline();
-                    lp->print_srcline_always_ = get_logger()->print_srcline_always();
-                    lp->level_ = cfgapi_table.logging.level;
-                    get_logger()->target_profiles()[(uint64_t)o] = lp;
-                    
-                }
-            }
-            //
-            if(CfgFactory::cfg_root()["settings"].lookupValue("sslkeylog_file", CfgFactory::get().sslkeylog_file_base)) {
-
-                CfgFactory::get().sslkeylog_file = CfgFactory::get().sslkeylog_file_base;
-
-                if(CfgFactory::get().sslkeylog_file.size() > 0) {
-
-                    CfgFactory::get().sslkeylog_file = string_format(CfgFactory::get().sslkeylog_file.c_str(),
-                                                                           CfgFactory::get().tenant_name.c_str());
-
-                    std::ofstream * o = new std::ofstream(CfgFactory::get().sslkeylog_file.c_str(),std::ios::app);
-                    get_logger()->targets(CfgFactory::get().sslkeylog_file,o);
-                    get_logger()->dup2_cout(false);
-                    get_logger()->level(cfgapi_table.logging.level);
-                    
-                    logger_profile* lp = new logger_profile();
-                    lp->print_srcline_ = get_logger()->print_srcline();
-                    lp->print_srcline_always_ = get_logger()->print_srcline_always();
-                    lp->level_ = loglevel(iINF,flag_add(iNOT,CRT|KEYS));
-                    get_logger()->target_profiles()[(uint64_t)o] = lp;
-                    
-                }
-            }
-            
-            
-            if( ! CfgFactory::get().syslog_server.empty() ) {
-                bool have_syslog = init_syslog();
-                if(! have_syslog) {
-                    ERRS_("syslog logging not set.");
-                }
-            }
-            
-            if(CfgFactory::cfg_root()["settings"].lookupValue("cfg_log_console", CfgFactory::get().log_console)) {
-                get_logger()->dup2_cout(CfgFactory::get().log_console);
-            }
-            
-/*            
- *          init_syslog();
-*/            
-        }
-    }
-    catch(const SettingNotFoundException &nfex) {
-    
-        FAT_("Setting not found: %s",nfex.getPath());
-        ret = false;
-    }
-    
-    return ret;
-}
 
 
 
@@ -463,7 +156,7 @@ int main(int argc, char *argv[]) {
                 get_logger()->dup2_cout(true);
                 
             case 'D':
-                cfg_daemonize = true;
+                SmithProxy::instance().cfg_daemonize = true;
                 break;
                 
             case 'i':
@@ -495,7 +188,7 @@ int main(int argc, char *argv[]) {
         set_logger(new QueueLogger());
     }
     
-    if(!cfg_daemonize) {
+    if(! SmithProxy::instance().cfg_daemonize) {
         std::thread* log_thread  = create_log_writer(get_logger());
         if(log_thread != nullptr) {
             pthread_setname_np(log_thread->native_handle(),string_format("sxy_lwr_%d",
@@ -547,7 +240,7 @@ int main(int argc, char *argv[]) {
     
     WARS_(" ");
     // set level to what's in the config
-    if (!load_config(CfgFactory::get().config_file)) {
+    if (! SmithProxy::instance().load_config(CfgFactory::get().config_file)) {
         if(CfgFactory::get().config_file_check_only) {
             FATS_("Config check: error loading config file.");
             exit(1);
@@ -595,7 +288,7 @@ int main(int argc, char *argv[]) {
         UDPCom::default_sock_family = AF_INET;
     }
     
-    if(cfg_daemonize) {
+    if(SmithProxy::instance().cfg_daemonize) {
         if(get_logger()->targets().size() <= 0) {
             FATS_("Cannot daemonize without logging to file.");
             exit(-5);
@@ -623,7 +316,7 @@ int main(int argc, char *argv[]) {
     
     //     atexit(__libc_freeres);   
 
-    if(cfg_openssl_mem_dbg) {
+    if(SmithProxy::instance().cfg_openssl_mem_dbg) {
 #ifndef USE_OPENSSL11
         CRYPTO_malloc_debug_init();
         CRYPTO_dbg_set_options(V_CRYPTO_MDEBUG_ALL);
@@ -657,7 +350,7 @@ int main(int argc, char *argv[]) {
 
     SSLCom::certstore()->destroy();
     
-    if(cfg_daemonize) {
+    if(SmithProxy::instance().cfg_daemonize) {
         this_daemon.unlink_pidfile();
     }
